Create Vaccine management system
As the name suggests, the vaccine management system is software that handles all the data related to vaccination. The data contains the name of the persons who have taken the vaccine and also tells the current status of vaccine availability. Previously the task of handling the vaccination data was very difficult, so there was a need for software that can handle all the vaccination data.